requirements.txt
absl-py==0.13.0
addict==2.4.0
albumentations==1.0.3
antlr4-python3-runtime==4.8
anyio==3.3.2
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.4b2
bleach==1.5.0
blessings==1.7
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
deprecation==2.1.0
detectron2==0.6+cu113
entrypoints==0.3
flatbuffers==2.0
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
grpcio==1.40.0
GuideConv==0.0.0
h5py==2.10.0
html5lib==0.9999999
hydra-core==1.1.1
idna==3.2
imageio==2.9.0
imgaug==0.4.0
importlib-resources==5.2.2
iopath==0.1.8
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyter-packaging==0.10.6
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mmcv-full==1.3.12
mypy-extensions==0.4.3
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nn-distance==0.0.0
nose==1.3.7
notebook==6.4.4
numpy==1.20.3
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pip==21.0.1
plyfile==0.7.4
portalocker==2.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.4.1
Send2Trash==1.8.0
setuptools==52.0.0.post20210125
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.8.30
toml==0.10.2
tomlkit==0.7.2
torch==1.10.1
torch-encoding==1.2.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.1
wheel==0.37.0
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.8
yapf==0.31.0
zipp==3.5.0
